+++ src/params.ml
@@ -1,5 +1,7 @@
 let debug = ref false
 
+let default_max_fuzzy_edit_distance = 3
+
 let preview_line_count = 5
 
 let default_max_file_tree_depth = 5